crawler/requirements.txt
scrapy
requests
python-dotenv